add ts-config and make project into lib
